Replace static assets & reorganize static directory
Significantly changes the `static` directory, with two major changes:

The `css`, `js`, and `img` subdirectories have been added, in order to organize static assets, and

Old assets have been removed and replaced with new assets for compatibility with template rewrites.

Small whitespace changes will be made to the new scripts in order to explain their functionality in the set of commits immediately following this.