Explain changes to `common.js`
This is an immaterial whitespace edit made to `buttons.js` in order to explain its changes in this commit summary.

Little has been changed about `common.js`, save for a new method `startDownload(downloadLink)` and accompanying variables.

The `startDownload` method takes a URL of a .zip file and writes it with the "magic phrase" prefix "http://!|" to the nwc24dl.bin file in the `/shared2/nwc24/` directory on the NAND. This is accomplished using the `addDownloadTask` function of the `wiiNwc24` plug-in, as well as the `launchTitle` function of the `ECommerceInterface` plug-in.