Create `base.j2`
This is the base Jinja2 template.

All other pages use this template as a base. It contains the necessary HTML tags to make the Shop Channel browser happy, and includes several scripts:

`jquery-1.3.2` - the last version of jQuery supported by Opera 9

`common.js`, `buttons.js`, and `tinycolor.js` - explained in earlier commits

Three blocks can be "filled in" by child templates that use this as a base. They are a style block, a script block, an onload script block, a page title block, and a content block.