Begin revamp: initial structure, pages, etc
It has been decided to work on a design more faithful and accurate to
the original Wii Shop Channel, as this is what many people will expect
and wish to have (including myself).

Assets have and will be recreated, including the buttons, which uses
SVGs as the source as to make it much easier to draft designs with as
they are responsive (however, they should be rendered to images once
finished due to numerous drawbacks to using them everywhere).
(We can't do fancy stuff with CSS here unfortunately,
the browser is too old!!)

This commit contains the initial structure, an accurate connecting
page and the very beginnings of the landing page.

All other old pages have been removed as they are yet to be rewritten.