Compatible controllers page
Well, I'm yet to make more of the badges and make it so they line up
horizontally with the right spacing.

I suspect the way templates work for this page will have to be done
differently, as this is shown by pressing the compatible controllers
button as well as when you're about to download a title, with title
downloading having a different header and an OK button.