Title search fixes
- Fix text box alignment
- Make title search keyboard hint translatable
- Move JavaScript and CSS out from being inline (and other consistency
  changes)
- Additionally, increase osc.py timeout to 2 seconds (but for me it
  still sometimes times out on launch, arghhh)

The button is still missing its outline when it's hovered though, but
that's going to be a bit of a pain to fix. In fact, anything related to
this is going to be a pain. I tried visiting numerous different ways of
approaching this and all the ones I tried that would have made this
better and just work didn't work on Opera on the Wii or didn't work on
Opera 9 at all because of how old it is. (God am I thankful for newer
web standards, as a good thing that would've been a tad hacky but would
have just worked completely fine had we had the ability to it is to
just put the btn class on the input, and put the placeholder attribute
on it for the "Search by Title" text, and then just set the placeholder
text to black. It would've been as simple as that, but with a few
changes to the input listener code between desktop and Wii.)

</rant>